route_finder/server/requirements.txt
absl-py==2.1.0 ; python_version >= "3.12" and python_version < "4.0"
annotated-types==0.7.0 ; python_version >= "3.12" and python_version < "4.0"
blinker==1.9.0 ; python_version >= "3.12" and python_version < "4.0"
cachetools==5.5.0 ; python_version >= "3.12" and python_version < "4.0"
certifi==2024.8.30 ; python_version >= "3.12" and python_version < "4.0"
charset-normalizer==3.4.0 ; python_version >= "3.12" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.12" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4.0" and platform_system == "Windows"
flask-cors==5.0.0 ; python_version >= "3.12" and python_version < "4.0"
flask-sqlalchemy==3.1.1 ; python_version >= "3.12" and python_version < "4.0"
flask==3.1.0 ; python_version >= "3.12" and python_version < "4.0"
geographiclib==1.52 ; python_version >= "3.12" and python_version < "4.0"
geojson-length==0.4.0 ; python_version >= "3.12" and python_version < "4.0"
geopy==2.2.0 ; python_version >= "3.12" and python_version < "4.0"
google-ai-generativelanguage==0.6.10 ; python_version >= "3.12" and python_version < "4.0"
google-api-core==2.23.0 ; python_version >= "3.12" and python_version < "4.0"
google-api-core[grpc]==2.23.0 ; python_version >= "3.12" and python_version < "4.0"
google-api-python-client==2.154.0 ; python_version >= "3.12" and python_version < "4.0"
google-auth-httplib2==0.2.0 ; python_version >= "3.12" and python_version < "4.0"
google-auth==2.36.0 ; python_version >= "3.12" and python_version < "4.0"
google-generativeai==0.8.3 ; python_version >= "3.12" and python_version < "4.0"
googleapis-common-protos==1.66.0 ; python_version >= "3.12" and python_version < "4.0"
greenlet==3.1.1 ; python_version < "3.13" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and python_version >= "3.12"
grpcio-status==1.68.0 ; python_version >= "3.12" and python_version < "4.0"
grpcio==1.68.0 ; python_version >= "3.12" and python_version < "4.0"
gunicorn==23.0.0 ; python_version >= "3.12" and python_version < "4.0"
httplib2==0.22.0 ; python_version >= "3.12" and python_version < "4.0"
idna==3.10 ; python_version >= "3.12" and python_version < "4.0"
itsdangerous==2.2.0 ; python_version >= "3.12" and python_version < "4.0"
jinja2==3.1.4 ; python_version >= "3.12" and python_version < "4.0"
markupsafe==3.0.2 ; python_version >= "3.12" and python_version < "4.0"
packaging==24.2 ; python_version >= "3.12" and python_version < "4.0"
proto-plus==1.25.0 ; python_version >= "3.12" and python_version < "4.0"
protobuf==5.28.3 ; python_version >= "3.12" and python_version < "4.0"
psycopg2==2.9.10 ; python_version >= "3.12" and python_version < "4.0"
pyasn1-modules==0.4.1 ; python_version >= "3.12" and python_version < "4.0"
pyasn1==0.6.1 ; python_version >= "3.12" and python_version < "4.0"
pydantic-core==2.27.1 ; python_version >= "3.12" and python_version < "4.0"
pydantic==2.10.1 ; python_version >= "3.12" and python_version < "4.0"
pyparsing==3.2.0 ; python_version >= "3.12" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.12" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.12" and python_version < "4"
sqlalchemy==2.0.36 ; python_version >= "3.12" and python_version < "4.0"
tqdm==4.67.0 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.12" and python_version < "4.0"
uritemplate==4.1.1 ; python_version >= "3.12" and python_version < "4.0"
urllib3==2.2.3 ; python_version >= "3.12" and python_version < "4.0"
werkzeug==3.1.3 ; python_version >= "3.12" and python_version < "4.0"
